fixed bugs

+++ StockApp/Models/TransactionViewModel.cs
@@ -3,8 +3,8 @@
 public class TransactionViewModel
 {
     public string Id { get; set; } = string.Empty;
-    public double Amount { get; set; }
-    public double PricePerUnit { get; set; }
+    public decimal Amount { get; set; }
+    public decimal PricePerUnit { get; set; }
     public DateTime Date { get; set; }
     public string Ticker { get; set; }
     public bool IsBuy { get; set; } = true;
